20_sprocs_evaluate_num_of_calls_last_hour.sql
This functions is looking for sprocs with a fixed number of calls (per hour, per day), which were called more/less times in the past hour. The idea is to catch missing scheduled jobs and/or jobs that got more aggressive (which may represent a problem as well...). As with other checks, the method is to compare the number of calls to past weeks.